+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ### BUG FIXES
 
 - [\#748](https://github.com/cosmos/evm/pull/748) Fix DynamicFeeChecker in Cosmos ante handler to respect NoBaseFee feemarkets' parameter.
+- [#19](https://github.com/zeta-chain/evm/pull/19) - Extract numeric chain ID before trying parse to UINT for eip712 transactions
 
 ## v0.5.0